refactor(rsc): convert admin system status page to server component
Fetches the initial health check server-side via a new src/lib/data/health.ts helper, then renders the live dashboard with a client StatusDashboard that keeps the 30s auto-refresh and manual refresh button.
